requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==25.0
pybit==5.9.0
pycryptodome==3.21.0
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
